refactor: remove thin-facade and consolidate types into relay-transport
- Move thin-facade type definitions to relay-transport.ts
- Update imports in relay-jsonrpc-transport.ts
- Delete unused thin-facade.ts and thin-facade.test.ts
- Following Hatago philosophy: 'Don't add, remove'

This simplifies the transport layer by removing an unnecessary abstraction
layer that was never implemented.
