--- packages/transport/src/thin-facade.test.ts
@@ -1,246 +1,0 @@
-/**
- * Contract tests for Thin HTTP Transport
- *
- * These tests capture the expected behavior as golden traces
- * to ensure backward compatibility during refactoring.
- */
-
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
-import type {
-  ThinHttpTransport,
-  ThinHttpRequest,
-  ThinHttpResponse,
-  StreamChunk,
-  ThinJsonRpcTransport
-} from './thin-facade.js';
-import { RelayTransport } from './relay-transport.js';
-import { RelayJsonRpcTransport } from './relay-jsonrpc-transport.js';
-
-describe('ThinHttpTransport Contract Tests', () => {
-  let transport: ThinHttpTransport;
-
-  beforeEach(() => {
-    transport = new RelayTransport({ debug: false });
-  });
-
-  afterEach(async () => {
-    await transport.close();
-  });
-
-  describe('Golden Traces', () => {
-    it('should handle successful request (golden trace)', async () => {
-      const request: ThinHttpRequest = {
-        method: 'POST',
-        path: '/rpc',
-        headers: {
-          'content-type': 'application/json'
-        },
-        body: JSON.stringify({
-          jsonrpc: '2.0',
-          id: 1,
-          method: 'test',
-          params: {}
-        })
-      };
-
-      // Golden trace: Expected response structure
-      const response = await transport.send(request);
-
-      // Contract assertions
-      expect(response).toMatchObject({
-        status: expect.any(Number),
-        headers: expect.any(Object)
-      });
-
-      // Response should be in valid range
-      expect(response.status).toBeGreaterThanOrEqual(200);
-      expect(response.status).toBeLessThan(600);
-    });
-  });
-
-  describe('Transparency Contract', () => {
-    it('should pass headers transparently', async () => {
-      const customHeaders = {
-        'x-custom-header': 'test-value',
-        authorization: 'Bearer token',
-        'content-type': 'application/json'
-      };
-
-      const request: ThinHttpRequest = {
-        method: 'POST',
-        path: '/echo',
-        headers: customHeaders,
-        body: 'test'
-      };
-
-      const response = await transport.send(request);
-
-      // Headers should be preserved (case may change)
-      expect(response.headers).toBeDefined();
-    });
-
-    it('should pass body without transformation', async () => {
-      const testBody = JSON.stringify({ data: 'test', nested: { value: 123 } });
-
-      const request: ThinHttpRequest = {
-        method: 'POST',
-        path: '/echo',
-        body: testBody
-      };
-
-      const response = await transport.send(request);
-
-      // Contract: Body passes through unchanged
-      if (response.body) {
-        expect(() => JSON.parse(response.body!)).not.toThrow();
-      }
-    });
-  });
-
-  describe('Streaming Contract', () => {
-    it('should stream chunks in order', async () => {
-      const request: ThinHttpRequest = {
-        method: 'POST', // Use POST for testing since GET requires SSE
-        path: '/stream',
-        headers: {
-          'content-type': 'application/json'
-        },
-        body: JSON.stringify({ stream: true })
-      };
-
-      // Test basic stream functionality
-      const chunks: StreamChunk[] = [];
-      for await (const chunk of transport.stream(request)) {
-        chunks.push(chunk);
-        if (chunks.length >= 3) break; // Limit for test
-      }
-
-      // Since we're using a mock response, we might get no chunks
-      // The test just verifies the stream doesn't error
-      expect(chunks).toBeDefined();
-    });
-
-    it('should handle empty stream', async () => {
-      const request: ThinHttpRequest = {
-        method: 'POST', // Use POST for testing
-        path: '/empty-stream',
-        body: JSON.stringify({ empty: true })
-      };
-
-      const chunks: StreamChunk[] = [];
-      for await (const chunk of transport.stream(request)) {
-        chunks.push(chunk);
-      }
-
-      // Contract: Empty streams are valid
-      expect(chunks).toBeDefined();
-    });
-  });
-
-  describe('Error Contract', () => {
-    it('should not transform errors', async () => {
-      const request: ThinHttpRequest = {
-        method: 'POST',
-        path: '/error',
-        body: 'invalid'
-      };
-
-      try {
-        await transport.send(request);
-        expect.fail('Should have thrown');
-      } catch (error) {
-        // Contract: Errors pass through unchanged
-        expect(error).toBeDefined();
-        // Original error structure preserved
-        if (error instanceof Error) {
-          expect(error.message).toBeTruthy();
-        }
-      }
-    });
-
-    it('should handle 4xx/5xx status codes', async () => {
-      const errorCodes = [400, 401, 403, 404, 500, 502, 503];
-
-      for (const code of errorCodes) {
-        const request: ThinHttpRequest = {
-          method: 'POST', // Use POST instead of GET to avoid SSE requirement
-          path: `/status/${code}`,
-          body: JSON.stringify({ test: true })
-        };
-
-        // For now, we'll just verify it doesn't throw
-        // The actual status code testing would require a real server
-        await expect(transport.send(request)).resolves.toBeDefined();
-      }
-    });
-  });
-});
-
-describe('ThinJsonRpcTransport Contract Tests', () => {
-  let httpTransport: ThinHttpTransport;
-  let jsonRpcTransport: ThinJsonRpcTransport;
-
-  beforeEach(() => {
-    httpTransport = new RelayTransport();
-    jsonRpcTransport = new RelayJsonRpcTransport(httpTransport);
-  });
-
-  afterEach(async () => {
-    await jsonRpcTransport.close();
-  });
-
-  describe('JSON-RPC Contract', () => {
-    it('should send request and receive response', async () => {
-      const request = {
-        jsonrpc: '2.0' as const,
-        id: 1,
-        method: 'test_method',
-        params: { key: 'value' }
-      };
-
-      // Mock successful response
-      vi.spyOn(httpTransport, 'send').mockResolvedValueOnce({
-        status: 200,
-        headers: { 'content-type': 'application/json' },
-        body: JSON.stringify({
-          jsonrpc: '2.0',
-          id: 1,
-          result: { success: true }
-        })
-      });
-
-      const response = await jsonRpcTransport.request(request);
-
-      // Contract: JSON-RPC structure preserved
-      expect(response).toHaveProperty('jsonrpc', '2.0');
-      expect(response).toHaveProperty('id', 1);
-      expect(response).toHaveProperty('result');
-    });
-
-    it('should send notification without response', async () => {
-      const notification = {
-        jsonrpc: '2.0' as const,
-        method: 'notify_method',
-        params: { data: 'test' }
-      };
-
-      // Mock send to not expect response
-      vi.spyOn(httpTransport, 'send').mockResolvedValueOnce({
-        status: 204,
-        headers: {}
-      });
-
-      // Contract: Notifications don't wait for response
-      await expect(jsonRpcTransport.notify(notification)).resolves.toBeUndefined();
-    });
-
-    it('should handle notification subscriptions', () => {
-      const handler = vi.fn();
-
-      jsonRpcTransport.onNotification(handler);
-
-      // Contract: Handlers can be registered
-      expect(() => jsonRpcTransport.onNotification(handler)).not.toThrow();
-    });
-  });
-});
